--- e2e/meta_test/test_schema_nodes_labels/test.yml
@@ -1,8 +1,0 @@
-query: >
-    CALL meta.schema() YIELD nodes, relationships UNWIND nodes AS node UNWIND node.labels AS label RETURN label ORDER BY label DESC;
-output:
-  - label: "University"
-  - label: "Student"
-  - label: "Person"
-  - label: "Person"
-  - label: "City"
