--- e2e/meta_test/test_schema_relationships_count/test.yml
@@ -1,7 +1,0 @@
-query: >
-    CALL meta.schema() YIELD nodes, relationships UNWIND relationships AS relationship RETURN relationship.properties.count AS count ORDER BY relationship.id DESC;
-output:
-  - count: 1
-  - count: 2
-  - count: 1
-  - count: 1
