--- e2e/meta_test/test_schema_relationships_labels/test.yml
@@ -1,7 +1,0 @@
-query: >
-    CALL meta.schema() YIELD nodes, relationships UNWIND relationships AS relationship RETURN relationship.label AS label ORDER BY relationship.id DESC;
-output:
-  - label: "LIVES_IN"
-  - label: "STUDIES_AT"
-  - label: "LIVES_IN"
-  - label: "IS_FRIENDS_WITH"
