+++ e2e/meta_util_test/test_schema_empty/test.yml
@@ -1,4 +1,4 @@
 query: >
-  CALL meta.schema() YIELD * RETURN *;
+  CALL meta_util.schema() YIELD * RETURN *;
 exception:
     - Can't generate a graph schema since there is no data in the database.
