+++ e2e/meta_util_test/test_schema_nodes_size/test.yml
@@ -1,0 +1,4 @@
+query: >
+    CALL meta_util.schema() YIELD nodes, relationships RETURN size(nodes) AS size;
+output:
+  - size: 4
