+++ e2e/meta_util_test/test_schema_relationships_count/test.yml
@@ -1,0 +1,7 @@
+query: >
+    CALL meta_util.schema() YIELD nodes, relationships UNWIND relationships AS relationship RETURN relationship.properties.count AS count ORDER BY relationship.id DESC;
+output:
+  - count: 1
+  - count: 2
+  - count: 1
+  - count: 1
