+++ e2e/meta_util_test/test_schema_relationships_size/test.yml
@@ -1,0 +1,4 @@
+query: >
+    CALL meta_util.schema() YIELD nodes, relationships RETURN size(relationships) AS size;
+output:
+  - size: 4
