--- python/mage/meta/__init__.py
@@ -1,1 +1,0 @@
-from mage.meta.parameters import Parameter  # noqa: F401, F402, F403
