+++ python/mage/meta_util/__init__.py
@@ -1,0 +1,1 @@
+from mage.meta_util.parameters import Parameter  # noqa: F401, F402, F403
